move functions into helper
