+++ src/utils/helpers.ts
@@ -1,5 +1,6 @@
 import '@formatjs/intl-numberformat/polyfill'
 import '@formatjs/intl-numberformat/locale-data/en'
+import ISO6391 from 'iso-639-1'
 import type { Facet, WorkMetadata } from 'src/data/types'
 import type { HorizontalBarRecord } from 'src/components/HorizontalStackedBarChart/HorizontalStackedBarChart'
 
@@ -97,5 +98,95 @@ export function toBarRecord(data: Facet) {
 }
 
 
+
+
+export function extractFOS(subjects: any) {
+  const fos = subjects
+    .filter((s) => s.subject.startsWith('FOS: '))
+    .map(({ subject: s }) => ({ id: kebabify(s.slice(5)), name: s.slice(5) }))
+
+  const uniqueFOS = Array.from(new Set(fos.map((f) => f.id))).map((id) =>
+    fos.find((f) => f.id === id)
+  )
+  return uniqueFOS
+}
+
+export function mapPeople(people: any[]) {
+  return people.map((p) => {
+    return {
+      ...p,
+      affiliation: p.affiliation.map((a) => ({
+        ...a,
+        id: a.affiliationIdentifier
+      })),
+      id: p.nameIdentifiers[0]?.nameIdentifier || ''
+    }
+  })
+}
+
+
+function kebabify(input: string): string {
+  return input
+    .replace(/([a-z])([A-Z])/g, '$1-$2') // Insert a dash between lowercase and uppercase letters
+    .toLowerCase()
+}
+
+
+
+interface Repository {
+  id: string
+  attributes: {
+    name: string
+  }
+}
+
+
+
+export function mapJsonToWork(json: any, included: any[]) {
+  const attrs = json.attributes
+  const repo =
+    json.relationships.client.data && included
+      ? included.find(
+        (r: Repository) => r.id === json.relationships.client.data.id
+      ) || null
+      : null
+
+  return {
+    ...attrs,
+    id: ID_BASE + json.id,
+    language: { id: attrs.language, name: ISO6391.getName(attrs.language) },
+    rights: attrs.rightsList,
+    creators: mapPeople(attrs.creators),
+    contributors: mapPeople(attrs.contributors),
+    fieldsOfScience: extractFOS(attrs.subjects),
+    registrationAgency: {
+      id: attrs.agency,
+      name: REGISTRATION_AGENCIES[attrs.agency]
+    },
+    repository: repo
+      ? { id: repo.id, name: repo.attributes.name }
+      : { id: '', name: '' },
+    schemaOrg: ''
+  }
+}
+
+
+
+const ID_BASE =
+  process.env.ENV === 'PROD'
+    ? 'https://doi.org/'
+    : 'https://handle.stage.datacite.org/'
+
+const REGISTRATION_AGENCIES = {
+  airiti: 'Airiti',
+  cnki: 'CNKI',
+  crossref: 'Crossref',
+  datacite: 'DataCite',
+  istic: 'ISTIC',
+  jalc: 'JaLC',
+  kisti: 'KISTI',
+  medra: 'mEDRA',
+  op: 'OP'
+}
